backend/requirements.txt
django
djoser
Pillow
psycopg2
django-filter
django-cleanup
psycopg2-binary
python-decouple
django-cors-headers
djangorestframework
djangorestframework-simplejwt